Feat/test: Moving features/tests from turn to board
